+++ rotkehlchen/api/rest.py
@@ -3523,24 +3523,24 @@ class RestAPI:
             is_dummy_pot=True,
         )
         if group_by_event_ids is True:
-            missing_accounting_rules = query_missing_accounting_rules(
+            event_accounting_rule_statuses = query_missing_accounting_rules(
                 db=self.rotkehlchen.data.db,
                 accounting_pot=accountant_pot,
                 evm_accounting_aggregator=accountant_pot.events_accountant.evm_accounting_aggregators,
                 events=[x for _, x in events_result],  # type: ignore
                 accountant=self.rotkehlchen.accountant,
-            )  # length of missing_accounting_rules and events guaranteeed by function
+            )  # length of missing_accounting_rules and events guaranteed by function
             entries = [  # type: ignore  # mypy doesn't understand significance of boolean check
                 x.serialize_for_api(  # type: ignore
                     customized_event_ids=customized_event_ids,
                     ignored_ids_mapping=ignored_ids_mapping,
                     hidden_event_ids=hidden_event_ids,
-                    missing_accounting_rule=missing_accounting_rule,
+                    event_accounting_rule_status=event_accounting_rule_status,
                     grouped_events_num=grouped_events_num,  # type: ignore
-                ) for (grouped_events_num, x), missing_accounting_rule in zip(events_result, missing_accounting_rules, strict=True)  # noqa: E501
+                ) for (grouped_events_num, x), event_accounting_rule_status in zip(events_result, event_accounting_rule_statuses, strict=True)  # noqa: E501
             ]
         else:
-            missing_accounting_rules = query_missing_accounting_rules(
+            event_accounting_rule_statuses = query_missing_accounting_rules(
                 db=self.rotkehlchen.data.db,
                 accounting_pot=accountant_pot,
                 evm_accounting_aggregator=accountant_pot.events_accountant.evm_accounting_aggregators,
@@ -3552,8 +3552,8 @@ class RestAPI:
                     customized_event_ids=customized_event_ids,
                     ignored_ids_mapping=ignored_ids_mapping,
                     hidden_event_ids=hidden_event_ids,
-                    missing_accounting_rule=missing_accounting_rule,
-                ) for x, missing_accounting_rule in zip(events_result, missing_accounting_rules, strict=True)  # noqa: E501
+                    event_accounting_rule_status=event_accounting_rule_status,
+                ) for x, event_accounting_rule_status in zip(events_result, event_accounting_rule_statuses, strict=True)  # noqa: E501
             ]
         result = {
             'entries': entries,