+++ rotkehlchen/externalapis/etherscan.py
@@ -216,6 +216,7 @@ class Etherscan(ExternalServiceWithApiKey, metaclass=ABCMeta):
         - RemoteError if there are any problems with reaching Etherscan or if
         an unexpected response is returned
         """
+        result = None
         query_str = f'https://{self.prefix_url}{self.base_url}/api?module={module}&action={action}'
         if options:
             for name, value in options.items():
@@ -244,21 +245,17 @@ class Etherscan(ExternalServiceWithApiKey, metaclass=ABCMeta):
             try:
                 response = self.session.get(query_str, timeout=timeout)
             except requests.exceptions.RequestException as e:
-                if 'Max retries exceeded with url' not in str(e):
-                    raise RemoteError(f'{self.chain} Etherscan API request failed due to {e!s}') from e  # noqa: E501
+                raise RemoteError(f'{self.chain} Etherscan API request failed due to {e!s}') from e
 
-            if (
-                    response is None or  # max retries exceeded with url
-                    response.status_code == HTTPStatus.TOO_MANY_REQUESTS
-            ):
+            if response.status_code == HTTPStatus.TOO_MANY_REQUESTS:
                 if backoff >= backoff_limit:
                     raise RemoteError(
-                        f'Getting {self.chain} Etherscan max retries/too many requests '
-                        f'error even after we incrementally backed off',
+                        f'Getting {self.chain} Etherscan too many requests error '
+                        f'even after we incrementally backed off',
                     )
 
                 log.debug(
-                    f'Got 429 or max retries exceeded from {self.chain} etherscan. Will '
+                    f'Got too many requests error from {self.chain} etherscan. Will '
                     f'backoff for {backoff} seconds.',
                 )
                 gevent.sleep(backoff)