+++ docs/changelog.rst
@@ -7,6 +7,9 @@ Changelog
 * :feature:`7051` Users will be able to add a custom price directly from the asset overview page.
 * :feature:`4841` The rotki data directory is now more organized.
 * :bug:`7147` rotki should no longer query price multiple times for the same asset across different chains.
+* :bug:`-` Failure to resolve etherscan api or hitting maximum backoff after too many retries will now no longer raise an exception.
+* :bug:`-` Some coinbase trades imported via the API should no longer show negative amounts. If you see negative amounts, purge and repull coinbase data.
+* :bug:`-` Removing exchanges will also remove them from the setting of non syncing exchanges.
 
 * :release:`1.31.2 <2023-12-22>`
 * :bug:`-` Importing timestamp in CSV in scientific format should now work without errors.